Evercam Media v1.0.1 EvercamMedia.ArchiveController

ArchiveController Module is being used to do all operations for Archives.

Functions

create(conn, params)

delete(conn, params)

index(conn, params)

import EvercamMedia.ArchiveController, only: [index: 2]

Method is being used to enlist all list of archives.

iex> conn.assigns[:current_user]
iex> %Plug.Conn{adapter: {Plug.Adapters.Cowboy.Conn, :...},
assigns: %{current_user: %User{__meta__: #Ecto.Schema.Metadata<:loaded, "users">,
  access_tokens: #Ecto.Association.NotLoaded<association :access_tokens is not loaded>,
  api_id: "--", api_key: "--"...}

current user is within conn.

iex> ensure_camera_exists(camera, exid, conn)
"Camera 'exid' not found!"

This method will ensure if the camera exists or not and work accordingly.

iex> ensure_can_list(current_user, camera, conn)
:ok
"Unauthorized"

This method will check the rights of user for given camera

Archive
|> Archive.by_camera_id(camera.id)
|> Archive.with_status_if_given(status)
|> Archive.get_all_with_associations

These are all EvercamMedia.Archive model functions.

iex> render(conn, ArchiveView, "index.json", %{archives: archives})

This will render EvercamMedia.ArchiveView for index.json

pending_archives(conn, )

show(conn, params)

update(conn, params)