Evercam Media v1.0.1 EvercamMedia.Router.Helpers

Module with named helpers generated from EvercamMedia.Router.

Functions

archive_path(conn_or_endpoint, action)

archive_path(conn_or_endpoint, action, params)

archive_path(conn_or_endpoint, action, id, params)

archive_path(conn_or_endpoint, action, id, archive_id, params)

archive_url(conn_or_endpoint, action)

archive_url(conn_or_endpoint, action, params)

archive_url(conn_or_endpoint, action, id, params)

archive_url(conn_or_endpoint, action, id, archive_id, params)

camera_path(conn_or_endpoint, action)

camera_path(conn_or_endpoint, action, params)

camera_path(conn_or_endpoint, action, id, params)

camera_share_path(conn_or_endpoint, action, id)

camera_share_path(conn_or_endpoint, action, id, params)

camera_share_request_path(conn_or_endpoint, action, id)

camera_share_request_path(conn_or_endpoint, action, id, params)

camera_share_request_url(conn_or_endpoint, action, id)

camera_share_request_url(conn_or_endpoint, action, id, params)

camera_share_url(conn_or_endpoint, action, id)

camera_share_url(conn_or_endpoint, action, id, params)

camera_url(conn_or_endpoint, action)

camera_url(conn_or_endpoint, action, params)

camera_url(conn_or_endpoint, action, id, params)

cloud_recording_path(conn_or_endpoint, action, id)

cloud_recording_path(conn_or_endpoint, action, id, params)

cloud_recording_url(conn_or_endpoint, action, id)

cloud_recording_url(conn_or_endpoint, action, id, params)

log_path(conn_or_endpoint, action, id)

log_path(conn_or_endpoint, action, id, params)

log_url(conn_or_endpoint, action, id)

log_url(conn_or_endpoint, action, id, params)

motion_detection_path(conn_or_endpoint, action, id)

motion_detection_path(conn_or_endpoint, action, id, params)

motion_detection_url(conn_or_endpoint, action, id)

motion_detection_url(conn_or_endpoint, action, id, params)

onvif_path(conn_or_endpoint, action, service, operation)

onvif_path(conn_or_endpoint, action, service, operation, params)

onvif_url(conn_or_endpoint, action, service, operation)

onvif_url(conn_or_endpoint, action, service, operation, params)

onvifptz_path(conn_or_endpoint, action, id)

onvifptz_path(conn_or_endpoint, action, id, params)

onvifptz_path(conn_or_endpoint, action, id, preset_token, params)

onvifptz_url(conn_or_endpoint, action, id)

onvifptz_url(conn_or_endpoint, action, id, params)

onvifptz_url(conn_or_endpoint, action, id, preset_token, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

public_path(conn_or_endpoint, action)

public_path(conn_or_endpoint, action, params)

public_url(conn_or_endpoint, action)

public_url(conn_or_endpoint, action, params)

snapshot_path(conn_or_endpoint, action)

snapshot_path(conn_or_endpoint, action, params)

snapshot_path(conn_or_endpoint, action, id, params)

snapshot_path(conn_or_endpoint, action, id, timestamp, params)

snapshot_path(conn_or_endpoint, action, id, year, month, params)

snapshot_path(conn_or_endpoint, action, id, year, month, day, params)

snapshot_path(conn_or_endpoint, action, id, year, month, day, hour, params)

snapshot_url(conn_or_endpoint, action)

snapshot_url(conn_or_endpoint, action, params)

snapshot_url(conn_or_endpoint, action, id, params)

snapshot_url(conn_or_endpoint, action, id, timestamp, params)

snapshot_url(conn_or_endpoint, action, id, year, month, params)

snapshot_url(conn_or_endpoint, action, id, year, month, day, params)

snapshot_url(conn_or_endpoint, action, id, year, month, day, hour, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

stream_path(conn_or_endpoint, action)

stream_path(conn_or_endpoint, action, token)

stream_path(conn_or_endpoint, action, token, params)

stream_path(conn_or_endpoint, action, token, filename, params)

stream_url(conn_or_endpoint, action)

stream_url(conn_or_endpoint, action, token)

stream_url(conn_or_endpoint, action, token, params)

stream_url(conn_or_endpoint, action, token, filename, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, id, params)

vendor_model_path(conn_or_endpoint, action)

vendor_model_path(conn_or_endpoint, action, params)

vendor_model_path(conn_or_endpoint, action, id, params)

vendor_model_url(conn_or_endpoint, action)

vendor_model_url(conn_or_endpoint, action, params)

vendor_model_url(conn_or_endpoint, action, id, params)

vendor_path(conn_or_endpoint, action)

vendor_path(conn_or_endpoint, action, params)

vendor_path(conn_or_endpoint, action, id, params)

vendor_url(conn_or_endpoint, action)

vendor_url(conn_or_endpoint, action, params)

vendor_url(conn_or_endpoint, action, id, params)