Evercam Media v1.0.1 EvercamMedia.Snapshot.DBHandler

This module should ideally delegate all the updates to be made to the database on various events to another module.

Right now, this is a extracted and slightly modified from the previous version of worker.

These are the list of tasks for the db handler

Create an entry in the snapshots table for each retrived snapshots

Update the CameraActivity table whenever there is a change in the camera status

Update the status and last_polled_at values of Camera table

Update the thumbnail_url of the Camera table - This was done in the previous version and not now. This update can be avoided if thumbnails can be dynamically served.

Functions

broadcast_change_to_users(camera)

change_camera_status(camera, timestamp, status, error_code \\ nil)

log_camera_status(camera, bool, datetime, error_code)

update_camera_status(camera_exid, timestamp, status, error_code \\ "generic", error_weight \\ 0)