Evercam Media v1.0.1 EvercamMedia.Snapshot.Poller

Provides functions and workers for getting snapshots from the camera

Functions can be called from other places to get snapshots manually.

Functions

get_config(cam_server)

Get the configuration of the camera worker.

handle_call(msg, from, state)

Server callback for stopping camera poller

handle_info(arg1, state)

Take care of unknown messages which otherwise would trigger function clause mismatch error.

init(args)

Initialize the camera server

start_link(args)

Start a poller for camera worker.

start_timer(cam_server)

Restart the poller for the camera that takes snapshot in frequent interval as defined in the args passed to the camera server.

stop_timer(cam_server)

Stop the poller for the camera.

update_config(cam_server, config)

Update the configuration of the camera worker