Evercam Media v1.0.1 EvercamMedia.Snapshot.WorkerSupervisor

This supervisor creates EvercamMedia.Snapshot.Worker using the strategy :simple_one_for_one and can only handle one child type of children.

Since we want to dynamically create EvercamMedia.Snapshot.Worker for the cameras, other types of strategies in supervisor are not suitable.

When creating a new worker, the supervisor passes on a list of @event_handlers. @event_handlers are the handlers that wants to react to the events generated by EvercamMedia.Snapshot.Worker. These handlers are automatically added to the event manager for every created worker.

Functions

get_config(camera)

Given a camera, it returns a map of values required for starting a camera worker.

init(atom)

Callback implementation for c:Supervisor.init/1.

initiate_workers()

Start a workers for each camera in the database.

This function is intended to be called after the EvercamMedia.Snapshot.WorkerSupervisor is initiated.

start_link()

start_worker(camera)

Start camera worker

update_worker(worker, camera)

Reinitialize camera worker with new configuration