Evercam Media v1.0.1 EvercamMedia.ArchiveController

ArchiveController Module is being used to do all operations for Archives.

Functions

create(conn, params)

delete(conn, params)

index(conn, params)

GET "/v1/cameras/:id/archives", ArchiveController, :index

index method is being used to enlist all archives of a given camera.

This endpoint is going through auth. So It will work by sending an HTTP request with API_KEY and API_ID.

Example and Code Explanation

GET /v1/cameras/dexter/archives?api_key={Your_API_key}&api_id={Your_API_id}

The resulting object will be list which will contain all the archives of a given camera.

{ "archives":
  [
    { "to_date": 1472716717,
      "title": "A Day in Seattle",
      "status": "Failed",
      "requester_name": "Dexter Morgan",
      "requester_email": "dexter@morgan.com",
      "requested_by": "dexter",
      "public":false,
      "id":"adayin-xc52",
      "from_date":1472715517,
      "frames":0,
      "embed_time":false,
      "created_at":1473068359,
      "camera_id":"dexter"
    }
  ]
}

id is being referred to camera_id. Your API_KEY and API_ID are being authenticated in EvercamMedia.Auth module. current_user is the person whom API_KEY and API_ID are present in HTTP_REQUEST_URL

pending_archives(conn, )

show(conn, params)

update(conn, params)